+++ Entity/GalleryManager.php
@@ -10,77 +10,20 @@
  */
 namespace Sonata\MediaBundle\Entity;
 
-use Sonata\MediaBundle\Model\GalleryManager as AbstractGalleryManager;
+use Sonata\CoreBundle\Entity\DoctrineBaseManager;
 use Sonata\MediaBundle\Model\GalleryInterface;
-use Doctrine\ORM\EntityManager;
 
-class GalleryManager extends AbstractGalleryManager
+class GalleryManager extends DoctrineBaseManager
 {
-    protected $em;
-    protected $repository;
-    protected $class;
-
     /**
-     * @param \Doctrine\ORM\EntityManager $em
-     * @param string                      $class
-     */
-    public function __construct(EntityManager $em, $class)
-    {
-        $this->em    = $em;
-        $this->class = $class;
-    }
-
-    /**
-     * @return mixed
-     */
-    protected function getRepository()
-    {
-        if (!$this->repository) {
-            $this->repository = $this->em->getRepository($this->class);
-        }
-
-        return $this->repository;
-    }
-
-    /**
-     * {@inheritdoc}
+     * BC Compatibility.
+     *
+     * @deprecated Please use save() from now
+     *
+     * @param GalleryInterface $gallery
      */
     public function update(GalleryInterface $gallery)
     {
-        $this->em->persist($gallery);
-        $this->em->flush();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getClass()
-    {
-        return $this->class;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function findOneBy(array $criteria)
-    {
-        return $this->getRepository()->findOneBy($criteria);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function findBy(array $criteria)
-    {
-        return $this->getRepository()->findBy($criteria);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function delete(GalleryInterface $gallery)
-    {
-        $this->em->remove($gallery);
-        $this->em->flush();
+        parent::save($gallery);
     }
 }